refactor: move helper funcs to n4htools
